backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pydantic==2.11.7
